Handle the ftcms custom scheme
This mirrors the behaviour of the original image service, first
attempting to load an image from the V1 API and then trying V2.

In order to utilise the proxying behaviour of this service we
don't load the image, we make a HEAD request to verify that an
image is available.

If neither can be found then we error.