Remove sentence about requesting an purging API key (#320)
We do not want to give users the ability to purge images themselves as it can be abused. Ideally they have an image publishing workflow that doesn't involve purging images. If an image needs to be purged, users can email or Slack the Origami team with a request to purge.