Whitelist the Origami image sets (#325)
We currently use xlink:href in flags to repeat the same path multiple
times. This now gets stripped, and it'd be a lot of work to update each
flag image. I've whitelisted all of the images in our image sets, which
means we should be vigilant when accepting new SVGs into our sets.