store images in Cloudinary using content hash as their ID (#620)
This changes the Origami Image Service's use of Cloudinary's fetch API with their upload API and changes how the images are stored.

Previously we used the image's url as the public ID when storing on Cloudinary. This is now causing issues because the publishing platforms which use the Origami Image Service have begun to use new image urls for the same image, meaning we have thousands upon thousands of duplicate images being stored in Cloudinary. _In a short manual scan over a few minutes I found over four thousand duplicate images._

This commit makes the Origami Image Service download the image first and calculate the SHA256 of the image, using the SHA value as the public ID. This solves the issue of duplicate images because the exact same image served at different URLs will now resolve to the same public ID in Cloudinary.

To set public ID's in Cloudinary is not possible using the fetch API which is why this commit also switches to the upload API.

This change does mean that the Origami Image Service is now downloading the full image before sending it to Cloudinary, which means it will have a slower response time and larger memory usage, this is only an issue for the first time an image is requested though because of all the caching in the Fastly CDN and Cloudinary's CDN, as well as the in-memory cache in the application code.